Added image drawing.
This was the initial goal of the new connector. Instead of having all
sorts of functions that draw directly on a "device", to have a graphic
library draw on an image and then draw the image on the device.

This could only be possible if images could have transparency. This is
the main cost to pay. Now each pixel is 2 bits instead of 1. I don't
think the cost is too expensive for the cleanliness of the results.